model_api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ayman Alkhawaleda <ayman.khawaleda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
pandas = "^2.2.1"
joblib = "^1.3.2"
scikit-learn = "1.2.2"
drf-spectacular = "^0.27.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
